Refactor script arguments and parameters
I have refactored the way the script handles arguments and parameters in order to make them more extensible and reilable.

The script now supports arguments with parameters, which can be seen with the "-f|--file" flag. 

The script also does not execute any flashing if the factory image file path has not been specified.